Team_40_GenderClassification/requirements.txt
torch
torchvision 
pillow
numpy
